Move Policies under components directory
