+++ packages/keychain/src/components/connect/CreateSession.tsx
@@ -8,7 +8,7 @@ import { Policy } from "@cartridge/controller";
 import { Button } from "@chakra-ui/react";
 import { useCallback, useState } from "react";
 import { useConnection } from "hooks/connection";
-import { Policies } from "Policies";
+import { Policies } from "components/Policies";
 import { ErrorAlert } from "components/ErrorAlert";
 import { SessionConsent } from "components/connect";
